Revert move of UPIFlowType
